favorite

+++ app/src/main/java/com/finalproject/profitableshopping/view/authentication/fragments/SignUpFragment.kt
@@ -124,7 +124,7 @@ class SignUpFragment : Fragment() {
                     ).observe(
                         this,
                         Observer {
-                            AppSharedPreference.setFavoriteId(requireContext(), it.toInt())
+                            AppSharedPreference.setFavoriteId(requireContext(), it)
                         }
                     )
                 } else {
